+++ ios/Podfile
@@ -34,6 +34,8 @@ target 'sharewaves' do
   pod 'glog', :podspec => '../node_modules/react-native/third-party-podspecs/glog.podspec'
   pod 'Folly', :podspec => '../node_modules/react-native/third-party-podspecs/Folly.podspec'
 
+  pod 'react-native-camera', :path => '../node_modules/react-native-camera'
+
   target 'sharewavesTests' do
     inherit! :search_paths
     # Pods for testing
