+++ package.json
@@ -14,7 +14,12 @@
     "react-native": "0.61.4",
     "react-redux": "^7.1.3",
     "react-spring": "^8.0.27",
-    "redux": "^4.0.4"
+    "redux": "^4.0.4",
+    "react-native-camera": "^3.9.0",
+    "react-native-gesture-handler": "^1.5.0",
+    "react-navigation": "^4.0.10",
+    "react-navigation-stack": "^1.10.3",
+    "react-timeout": "^1.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.6.2",
